Fix invoker infinite loop after receiving signals
After receiving a signal, a pipe is written, which is polled by select.
As of ac1ed61, that pipe is never read, causing the select() to
repeatedly trigger and use 100% CPU after receiving any signal that
doesn't cause the process to terminate.

Partial revert of ac1ed61, plus a one-line fix for the select() fd set