[mapplauncherd] Remove applauncherd and refactor boosters into independent processes
Having a single applauncherd process with boosters as plugins is a
flawed idea: most notably, it's not possible to have two boosters with
conflicting libraries, like Qt4 and Qt5.

A much cleaner solution is to turn libapplauncherd into a support
library for boosters, and have each booster run as an independent
process.