fix: make deepin-turbo a proper shared library
The original approach defined deepin-turbo as a MODULE target but then
try to link it explicitly. I don't really understand what happened here
but it looks wrong.

Redefine it as a SHARED target allows linking to it with a proper
target_link_libraries without having to specify paths and dependencies
manually, which actually fails when using cmake -GNinja.

Change-Id: Iee818057ff484959b1fb0b4b52db674b1fc6ded2
Reviewed-on: http://gerrit.uniontech.com/c/deepin-turbo/+/1568
Reviewed-by: <mailman@uniontech.com>
Reviewed-by: zhangdingyuan <justforlxz@gmail.com>
Tested-by: <mailman@uniontech.com>